Prevent downloading .tar files others are downloading without -o param

+++ src/python/Poststorm_Imagery/collector/collect.py
@@ -160,13 +160,20 @@ if OPTIONS.download:
 
             while download_incomplete:
                 try:
-                    tar.download_url(output_dir=save_path, user=OPTIONS.user, overwrite=OPTIONS.overwrite)
-                    if TarRef.verify_integrity(tar.tar_file_path) is False:
-                        print('Integrity could not be verified!')
-                        os.remove(tar.tar_file_path)
+                    lock_info = helpers.get_lock_info(part_file=tar.tar_file_path + '.part')
+                    if type(lock_info['user']) != str or OPTIONS.user == lock_info['user']:
+
+                        tar.download_url(output_dir=save_path, user=OPTIONS.user, overwrite=OPTIONS.overwrite)
+                        if TarRef.verify_integrity(tar.tar_file_path) is False:
+                            print('Integrity could not be verified!')
+                            os.remove(tar.tar_file_path)
+                        else:
+                            print('Extracting files...')
+                            TarRef.extract_archive(tar.tar_file_path)
+                            download_incomplete = False
+
                     else:
-                        print('Extracting files...')
-                        TarRef.extract_archive(tar.tar_file_path)
+                        print('Another user is in the process of downloading ' + tar.tar_file_name + '.tar!  ... Skipping')
                         download_incomplete = False
                 except Exception as e:
                     if e == KeyboardInterrupt or SystemExit:
